requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
PySocks==1.7.1
python-telegram-bot==12.6.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
tweepy==3.9.0
urllib3==1.25.9
Werkzeug==1.0.1
